Refactor the data flow system
