+++ src/app/requests.service.ts
@@ -1,19 +1,8 @@
 import { Injectable } from '@angular/core';
 import { IRole, ILocation, IDevice, IActivity, IWidget } from './shared/Definitions';
 import { PermissionsService } from './permissions.service';
-import * as locationsMocks from './locations/locations.mock';
-import * as devicesMocks from './devices/devices.mock';
 import { sample, times} from 'lodash';
 import { MocksService } from './mocks.service';
-function makeMockWidget (device: IDevice, location: ILocation): IWidget {
-  const widget: IWidget = {
-    device: device,
-    location: location,
-    name : sample(['Cloud', 'Lamp', 'Roberry']),
-    pin: sample(device.pins)
-  };
-  return widget;
-}
 
 
 @Injectable()
@@ -22,14 +11,10 @@ export class RequestsService {
   constructor(private permissions: PermissionsService, private mocks: MocksService) { }
 
   async getWidgets (): Promise<Array<IWidget>> {
-    const devices = await this.getDevices();
-    const locations = await this.getLocations();
-    return times(4, () => {
-      return makeMockWidget(sample(devices), sample(locations));
-    });
+    return this.mocks.Widgets();
   }
   async getLocations(): Promise<Array<ILocation>> {
-    return locationsMocks.generateMock();
+    return this.mocks.Locations();
   }
   
   async getActivities (): Promise<Array<IActivity>> {
@@ -37,30 +22,11 @@ export class RequestsService {
   }
 
   async getDevices (): Promise<Array<IDevice>> {
-    return devicesMocks.generateMock();
+    return this.mocks.Devices();
   }
 
   async getRoles (): Promise<Array<IRole>> {
-    return [
-      {
-        permissions: [
-          this.permissions.findByKey('DEVICES_PIN_CHANGE'),
-          this.permissions.findByKey('UPDATE_USER')
-        ],
-        id: 1,
-        title: 'Content Moderator'
-      },
-      {
-        permissions: [],
-        id: 2,
-        title: 'Higher manager'
-      },
-      {
-        permissions: [],
-        id: 3,
-        title: 'Installler'
-      }
-    ];
+    return this.mocks.Roles();
   }
 
 }
